+++ src/main/kotlin/dsm/pick2024/domain/status/port/in/ChangeStatusUseCase.kt
@@ -1,6 +1,6 @@
 package dsm.pick2024.domain.status.port.`in`
 
-import dsm.pick2024.domain.status.present.dto.request.ChangeStatusRequest
+import dsm.pick2024.domain.status.presentation.dto.request.ChangeStatusRequest
 
 interface ChangeStatusUseCase {
     fun changeStatus(request: List<ChangeStatusRequest>)
