+++ src/main/kotlin/dsm/pick2024/domain/status/port/in/QueryClassStatusUseCase.kt
@@ -1,6 +1,6 @@
 package dsm.pick2024.domain.status.port.`in`
 
-import dsm.pick2024.domain.status.present.dto.response.QueryClassResponse
+import dsm.pick2024.domain.status.presentation.dto.response.QueryClassResponse
 
 interface QueryClassStatusUseCase {
     fun queryClasStatus(
